+++ src/MusicalScore/VoiceData/Note.ts
@@ -96,6 +96,8 @@ export class Note {
     private noteheadColor: string;
     private noteheadColorCurrentlyRendered: string;
     public Fingering: TechnicalInstruction; // this is also stored in VoiceEntry.TechnicalInstructions
+    public StringInstruction: TechnicalInstruction; // this is also stored in VoiceEntry.TechnicalInstructions
+    // note that there is also TabNote.StringNumber, so we can't use that identifier here
     /** Used by GraphicalNote.FromNote(note) and osmd.rules.GNote(note) to get a GraphicalNote from a Note. */
     public NoteToGraphicalNoteObjectId: number; // used with EngravingRules.NoteToGraphicalNoteMap